+++ Entities/AEmitter.cpp
@@ -171,7 +171,7 @@ int AEmitter::ReadProperty(std::string propName, Reader &reader)
         reader >> burstSize;
         // Go through all emissions and set the rate so that it emulates the way it used to work, for mod backwards compatibility
         for (list<Emission *>::iterator eItr = m_EmissionList.begin(); eItr != m_EmissionList.end(); ++eItr)
-            (*eItr)->m_BurstSize = ceilf((float)burstSize / (float)m_EmissionList.size());
+            (*eItr)->m_BurstSize = std::ceil((float)burstSize / (float)m_EmissionList.size());
     }
     else if (propName == "BurstScale")
         reader >> m_BurstScale;
@@ -577,7 +577,7 @@ void AEmitter::Update()
             // Don't set the flipping for the flash because that is wasting resources when drawing,
             // just handle the flipping of the rotation here.
             m_pFlash->SetRotAngle(m_HFlipped ? c_PI + m_Rotation.GetRadAngle() - m_EmitAngle.GetRadAngle() : m_Rotation.GetRadAngle() + m_EmitAngle.GetRadAngle());
-//            m_pFlash->SetFrame(floorf((m_pFlash->GetFrameCount()/* - 1*/) * PosRand() - 0.001));
+//            m_pFlash->SetFrame(std::floor((m_pFlash->GetFrameCount()/* - 1*/) * PosRand() - 0.001));
             m_pFlash->SetScale(m_FlashScale);
             m_pFlash->SetNextFrame();
             m_pFlash->Update();
