+++ Managers/FrameMan.cpp
@@ -662,17 +662,17 @@ namespace RTE {
 		// Just make the alt the same color as the main one if no one was specified
 		if (altColor == 0) { altColor = color; }
 
-		intPos[X] = floorf(start.m_X);
-		intPos[Y] = floorf(start.m_Y);
+		intPos[X] = std::floor(start.m_X);
+		intPos[Y] = std::floor(start.m_Y);
 
 		// Wrap line around the scene if it makes it shorter
 		if (shortestWrap) {
 			Vector deltaVec = g_SceneMan.ShortestDistance(start, end, false);
-			delta[X] = floorf(deltaVec.m_X);
-			delta[Y] = floorf(deltaVec.m_Y);
+			delta[X] = std::floor(deltaVec.m_X);
+			delta[Y] = std::floor(deltaVec.m_Y);
 		} else {
-			delta[X] = floorf(end.m_X) - intPos[X];
-			delta[Y] = floorf(end.m_Y) - intPos[Y];
+			delta[X] = std::floor(end.m_X) - intPos[X];
+			delta[Y] = std::floor(end.m_Y) - intPos[Y];
 		}
 		if (delta[X] == 0 && delta[Y] == 0) {
 			return 0;
